Trap added

+++ src/tacticode/spell/ISpell.hpp
@@ -14,7 +14,7 @@ namespace tacticode
 		{
 		protected:
 			std::string m_name;
-			size_t m_power;
+			float m_power;
 			size_t m_range;
 			size_t m_aoe;
 			size_t m_cooldown;
@@ -25,7 +25,7 @@ namespace tacticode
 			virtual ~ISpell() = default;
 			virtual void castSpell() = 0;
 			virtual const std::string& getName() const = 0;
-			virtual const size_t getPower() const = 0;
+			virtual const float getPower() const = 0;
 			virtual const size_t getRange() const = 0;
 			virtual const size_t getCooldown() const = 0;
 			virtual const size_t getIsActive() const = 0;
